chore: update theme logic
